Allow omitting an input file to v/m parser
Allow just one input file to the combined variants mutations parser,
doing away with the need to create dummy files.

From here, it would actually make a lot of sense to just allow any
number of files to be passed, and sort based on headers.
But that's a task for another time.